CODIGO_ROSTRO_IA/requirements.txt
streamlit
ultralytics
opencv-python-headless
pandas
numpy
Pillow